Move pre/post-processing functions and wrapper to lib_masked_loading.py
