+++ fractal_tasks_core/cellpose_segmentation.py
@@ -38,6 +38,7 @@ from cellpose.core import use_gpu
 import fractal_tasks_core
 from fractal_tasks_core.lib_channels import ChannelNotFoundError
 from fractal_tasks_core.lib_channels import get_channel_from_image_zarr
+from fractal_tasks_core.lib_masked_loading import wrap_function_with_masking
 from fractal_tasks_core.lib_pyramid_creation import build_pyramid
 from fractal_tasks_core.lib_regions_of_interest import (
     array_to_bounding_box_table,
@@ -47,228 +48,12 @@ from fractal_tasks_core.lib_regions_of_interest import (
 )
 from fractal_tasks_core.lib_ROI_overlaps import find_overlaps_in_ROI_indices
 from fractal_tasks_core.lib_ROI_overlaps import get_overlapping_pairs_3D
-from fractal_tasks_core.lib_upscale_array import convert_region_to_low_res
-from fractal_tasks_core.lib_upscale_array import upscale_array
 from fractal_tasks_core.lib_zattrs_utils import extract_zyx_pixel_sizes
 from fractal_tasks_core.lib_zattrs_utils import rescale_datasets
 
 logger = logging.getLogger(__name__)
 
 __OME_NGFF_VERSION__ = fractal_tasks_core.__OME_NGFF_VERSION__
-
-
-def preprocess_cellpose_input(
-    image_array: np.ndarray,
-    *,
-    region: tuple[slice] = None,
-    current_label_path: str = None,
-    ROI_table_path: str = None,
-    ROI_index: int = None,  # FIXME: naming?
-) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
-    """
-    Preprocess a four-dimensional cellpose input
-
-    This involves :
-
-    - Loading the masking label array for the appropriate region;
-    - Extracting the appropriate label value from the ``ROI_table_obs``
-      dataframe;
-    - Constructing the background mask, where the masking label matches with a
-      specific label value;
-    - Setting the background of ``image_array`` to ``0``;
-    - Loading the array which will be needed in postprocessing to restore
-      background.
-
-    **FIXME 1**: Review/improve variable names
-
-    **FIXME 2**: Make this function more flexible, and move it to a separate
-    module (see `issue 340
-    <https://github.com/fractal-analytics-platform/fractal-tasks-core/issues/340>`_).
-
-    **FIXME 3**: This function relies on a proposed change to OME-NGFF table
-    specs (https://github.com/ome/ngff/pull/64).
-
-    Current naming of variables refers to a two-steps labeling ("first identify
-    organoids, then look for nuclei inside each organoid") :
-
-    - "masking" refers to the labels that are used to identify the object vs
-      background (e.g. the organoid labels); these labels already exist.
-    - "current" refers to the labels that are currently being computed in the
-      `cellpose_segmentation` task, e.g. the nuclear labels.
-
-
-    :param image_array: 4D CZYX array with image data for a specific ROI.
-    :param region: The ZYX indices to be used, in a form like ``(slice(0, 1),
-                   slice(1000, 2000), slice(1000, 2000))``.
-    :param current_label_path: Path to the image used as current
-                               label, in a form like
-                               ``/somewhere/plate.zarr/A/01/0/labels/nuclei_in_organoids/0``.
-    :param ROI_table_obs: ``obs`` attribute of the AnnData table for the
-                          masking-label ROIs; this is used (together with
-                          ``ROI_index``) to extract ``label_value``.
-    :param ROI_index: index of the current ROI, which is used to extract
-                      ``label_value`` from ``ROI_table_obs``.
-    """
-
-    logger.info(f"[preprocess_cellpose_input] {image_array.shape=}")
-    logger.info(f"[preprocess_cellpose_input] {region=}")
-
-    # Check that image data are 4D (CZYX) - FIXME issue 340
-    if not image_array.ndim == 4:
-        raise ValueError(
-            "preprocess_cellpose_input requires a 4D "
-            f"image_array argument, but {image_array.shape=}"
-        )
-
-    # Load the ROI table and its metadata attributes
-    ROI_table = ad.read_zarr(ROI_table_path)
-    ROI_table_obs = ROI_table.obs
-    attrs = zarr.group(ROI_table_path).attrs
-    if not attrs["type"] == "ngff:region_table":
-        raise ValueError("Wrong attributes for {ROI_table_path}:\n{attrs}")
-    label_relative_path = attrs["region"]["path"]
-    column_name = attrs["instance_key"]
-
-    # Check that ROI_table_obs has the right column and extract label_value
-    if column_name not in ROI_table_obs.columns:
-        raise ValueError(
-            'In preprocess_cellpose_input, "{column_name}" '
-            f" missing in {ROI_table_obs.columns=}"
-        )
-    label_value = int(ROI_table_obs[column_name][ROI_index])
-
-    # Load masking-label array (lazily)
-    masking_label_path = str(
-        Path(ROI_table_path).parent / label_relative_path / "0"
-    )
-    logger.critical(f"{masking_label_path=}")
-    masking_label_array = da.from_zarr(masking_label_path)
-    logger.info(
-        f"[preprocess_cellpose_input] {masking_label_path=}, "
-        f"{masking_label_array.shape=}"
-    )
-
-    # Load current-label array (lazily)
-    current_label_array = da.from_zarr(current_label_path)
-    logger.info(
-        f"[preprocess_cellpose_input] {current_label_path=}, "
-        f"{current_label_array.shape=}"
-    )
-
-    # Load ROI data for current label array
-    current_label_region = current_label_array[region].compute()
-
-    # Load ROI data for masking label array, with or without upscaling
-    if masking_label_array.shape != current_label_array.shape:
-        logger.info("Upscaling of masking label is needed")
-        lowres_region = convert_region_to_low_res(
-            highres_region=region,
-            highres_shape=current_label_array.shape,
-            lowres_shape=masking_label_array.shape,
-        )
-        masking_label_region = masking_label_array[lowres_region].compute()
-        masking_label_region = upscale_array(
-            array=masking_label_region,
-            target_shape=current_label_region.shape,
-        )
-    else:
-        masking_label_region = masking_label_array[region].compute()
-
-    # Check that all shapes match
-    shapes = (
-        masking_label_region.shape,
-        current_label_region.shape,
-        image_array.shape[1:],
-    )
-    if len(set(shapes)) > 1:
-        raise ValueError(
-            "Shape mismatch:\n"
-            f"{current_label_region.shape=}\n"
-            f"{masking_label_region.shape=}\n"
-            f"{image_array.shape=}"
-        )
-
-    # Compute background mask
-    background_3D = masking_label_region != label_value
-    if (masking_label_region == label_value).sum() == 0:
-        raise ValueError(
-            f"Label {label_value} is not present in the extracted ROI"
-        )
-
-    # Set image background to zero
-    background_4D = np.expand_dims(background_3D, axis=0)
-    image_array[background_4D] = 0
-
-    return (image_array, background_3D, current_label_region)
-
-
-def postprocess_cellpose_output(
-    *,
-    modified_array: np.ndarray,
-    original_array: np.ndarray,
-    background: np.ndarray,
-) -> np.ndarray:
-    """
-    Postprocess a cellpose input, mainly to restore its original background
-
-    **FIXME 1**: review/improve variable names
-
-    **FIXME 2**: make this function more flexible, and move it to a separate
-    module (see `issue 340
-    <https://github.com/fractal-analytics-platform/fractal-tasks-core/issues/340>`_).
-
-    :param modified_array: The 3D (ZYX) array with the correct object data and
-                           wrong background data.
-    :param original_array: The 3D (ZYX) array with the wrong object data and
-                           correct background data.
-    :param background: The 3D (ZYX) boolean array that defines the
-                       background.
-    """
-
-    # Restore background
-    modified_array[background] = original_array[background]
-    return modified_array
-
-
-def wrap_function_with_masking(
-    *,
-    function: Callable,
-    image_array: np.ndarray,
-    kwargs: Optional[dict] = None,
-    use_masks: bool,
-    preprocessing_kwargs: Optional[dict] = None,
-    postprocessing_kwargs: Optional[dict] = None,
-):
-    """
-    Wrap a function with some pre/post-processing functions
-
-    :param function: The callable function to be wrapped.
-    :param args: Positional arguments for ``function``.
-    :param kwargs: Keyword arguments for ``function``.
-    :param use_masks: If ``False``, the wrapper only calls ``function(*args,
-                      **kwargs)``.
-    :param preprocessing_kwargs: Keyword arguments for the preprocessing
-                                 function.
-    :param postprocessing_kwargs: Keyword arguments for the postprocessing
-                                  function.
-    """
-    if use_masks:
-        (
-            image_array,
-            background_3D,
-            current_label_region,
-        ) = preprocess_cellpose_input(image_array, **preprocessing_kwargs)
-
-    new_label_img = function(image_array, **kwargs)
-
-    if use_masks:
-        new_label_img = postprocess_cellpose_output(
-            modified_array=new_label_img,
-            original_array=current_label_region,
-            background=background_3D,
-        )
-
-    return new_label_img
 
 
 def segment_ROI(
